pyproject.toml
[tool.poetry]
name = "fastapi-toolhub-oauth"
version = "0.1.0"
description = ""
authors = ["Slavina Stefanova <sstefanova@wikimedia.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.114.1"}
httpx = "^0.27.2"
itsdangerous = "^2.2.0"
jwt = "^1.3.1"
python-jose = "^3.3.0"
pydantic-settings = "^2.5.2"
cryptography = "^43.0.1"
tortoise-orm = "^0.21.6"
asyncmy = "^0.2.9"
aerich = "^0.7.2"
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"

[tool.ruff.lint]
extend-select = ["I"]

[tool.aerich]
tortoise_orm = "backend.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
